Adding: DSE - Reinstall macOS
This script is to be used in a JAMF policy. Paramater 4 requires the OS vanity name:
Catalina or Mojave.

The script will then do a reinstall of the macOS on the drive, whiping the contents
of said drive.

Requirements:

Requires that Install macOS Catalina.app or Install macOS Mojave.app be located on the computer
in the Applications folder. In our case the Macintosh installer is packaged up in a DMG with
Composer which Jamf will push out in the policy before executing this script.