Create SetupARD
Author: Andrew W. Johnson
Date: Spring 2020.
Version 2.00
Organization: Stony Brook Univeristy/DoIT

This script was originally written by Armin Briegel
https://scriptingosx.com/2016/01/control-apple-remote-desktop-access-with-munki/
Modified for our environment, and then to port it over to JAMF and ZSH.
Adds the two possible local accounts we use access the computer through ARD and allows 
all privileges. It then removes any other accounts from accessing this computer through
ARD and removes all privileges.

The script will also add the ard_admin group to the system and in that group it nests an
Ad group. I'm not sure this is entirely necessary.

Script is a bit clunky, still learning ZSH, and do not feel the as proficient with it, as
say Perl.